requirements.txt
transformers>=4.38.0
torch>=2.1.0
accelerate>=0.26.0
bitsandbytes>=0.42.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.9.0
pydicom>=2.4.4
pydantic>=2.5.0
python-jose>=3.3.0
aiofiles>=23.2.1
